move evict_range to RangeCacheExt
