+++ components/hybrid_engine/src/range_cache_engine.rs
@@ -1,6 +1,6 @@
 // Copyright 2024 TiKV Project Authors. Licensed under Apache-2.0.
 
-use engine_traits::{KvEngine, RangeCacheEngine, RangeCacheEngineExt};
+use engine_traits::{CacheRange, KvEngine, RangeCacheEngine, RangeCacheEngineExt};
 
 use crate::HybridEngine;
 
@@ -11,5 +11,10 @@ where
 {
     fn range_cache_engine_enabled(&self) -> bool {
         true
+    }
+
+    #[inline]
+    fn evict_range(&self, range: &CacheRange) {
+        self.region_cache_engine().evict_range(range);
     }
 }
